Move non-form properties and methods to base class
